+++ lib/src/internal/marshaller.dart
@@ -1,6 +1,7 @@
 import 'package:meta/meta.dart';
 
 import '../ffi/types.dart';
+import '../ffi/utf8.dart';
 import '../modules/exceptions.dart';
 import '../modules/memory.dart';
 import 'invoker_generated.dart';
@@ -110,6 +111,24 @@ T _toDartType<T>(Object o, Memory bind) {
       } else {
         throw MarshallingException.noAddress(o);
       }
+    } else if (T == Pointer_UintPtr) {
+      if (o is int) {
+        return Pointer<UintPtr>.fromAddress(o, bind) as T;
+      } else {
+        throw MarshallingException.noAddress(o);
+      }
+    } else if (T == Pointer_Bool) {
+      if (o is int) {
+        return Pointer<Bool>.fromAddress(o, bind) as T;
+      } else {
+        throw MarshallingException.noAddress(o);
+      }
+    } else if (T == Pointer_Int) {
+      if (o is int) {
+        return Pointer<Int>.fromAddress(o, bind) as T;
+      } else {
+        throw MarshallingException.noAddress(o);
+      }
     } else if (T == Pointer_Int8) {
       if (o is int) {
         return Pointer<Int8>.fromAddress(o, bind) as T;
@@ -137,6 +156,12 @@ T _toDartType<T>(Object o, Memory bind) {
     } else if (T == Pointer_Double) {
       if (o is int) {
         return Pointer<Double>.fromAddress(o, bind) as T;
+      } else {
+        throw MarshallingException.noAddress(o);
+      }
+    } else if (T == Pointer_UnsignedInt) {
+      if (o is int) {
+        return Pointer<UnsignedInt>.fromAddress(o, bind) as T;
       } else {
         throw MarshallingException.noAddress(o);
       }
@@ -176,6 +201,12 @@ T _toDartType<T>(Object o, Memory bind) {
       } else {
         throw MarshallingException.noAddress(o);
       }
+    } else if (T == Pointer_Utf8) {
+      if (o is int) {
+        return Pointer<Utf8>.fromAddress(o, bind) as T;
+      } else {
+        throw MarshallingException.noAddress(o);
+      }
     } else if (T == Pointer_Opaque) {
       if (o is int) {
         return Pointer<Opaque>.fromAddress(o, bind) as T;
@@ -198,6 +229,24 @@ T _toDartType<T>(Object o, Memory bind) {
       } else if (T == Pointer_Pointer_IntPtr) {
         if (o is int) {
           return Pointer<Pointer<IntPtr>>.fromAddress(o, bind) as T;
+        } else {
+          throw MarshallingException.noAddress(o);
+        }
+      } else if (T == Pointer_Pointer_UintPtr) {
+        if (o is int) {
+          return Pointer<Pointer<UintPtr>>.fromAddress(o, bind) as T;
+        } else {
+          throw MarshallingException.noAddress(o);
+        }
+      } else if (T == Pointer_Pointer_Bool) {
+        if (o is int) {
+          return Pointer<Pointer<Bool>>.fromAddress(o, bind) as T;
+        } else {
+          throw MarshallingException.noAddress(o);
+        }
+      } else if (T == Pointer_Pointer_Int) {
+        if (o is int) {
+          return Pointer<Pointer<Int>>.fromAddress(o, bind) as T;
         } else {
           throw MarshallingException.noAddress(o);
         }
@@ -231,6 +280,12 @@ T _toDartType<T>(Object o, Memory bind) {
         } else {
           throw MarshallingException.noAddress(o);
         }
+      } else if (T == Pointer_Pointer_UnsignedInt) {
+        if (o is int) {
+          return Pointer<Pointer<UnsignedInt>>.fromAddress(o, bind) as T;
+        } else {
+          throw MarshallingException.noAddress(o);
+        }
       } else if (T == Pointer_Pointer_Uint8) {
         if (o is int) {
           return Pointer<Pointer<Uint8>>.fromAddress(o, bind) as T;
@@ -258,6 +313,12 @@ T _toDartType<T>(Object o, Memory bind) {
       } else if (T == Pointer_Pointer_Char) {
         if (o is int) {
           return Pointer<Pointer<Char>>.fromAddress(o, bind) as T;
+        } else {
+          throw MarshallingException.noAddress(o);
+        }
+      } else if (T == Pointer_Pointer_Utf8) {
+        if (o is int) {
+          return Pointer<Pointer<Utf8>>.fromAddress(o, bind) as T;
         } else {
           throw MarshallingException.noAddress(o);
         }